requirements.txt
PyQt6>=6.4.0
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tifffile>=2023.4.12
imagecodecs>=2023.1.23
Wand>=0.6.13
python-dotenv>=1.0.0
pypdf>=3.0.0
tqdm>=4.67.0
tabulate>=0.9.0 
img2pdf>=0.4.5

